pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "grados"
version = "0.6.6"
description = "Academic research MCP server — search, extract, and manage papers"
readme = "README.md"
license = "MIT"
requires-python = ">=3.11"
authors = [{ name = "macfish" }]
keywords = ["mcp", "academic", "research", "papers", "pdf", "doi", "zotero", "sciencedirect"]
classifiers = [
    "Development Status :: 4 - Beta",
    "Intended Audience :: Science/Research",
    "Topic :: Scientific/Engineering",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Programming Language :: Python :: 3.13",
]

dependencies = [
    "fastmcp >= 2.0",
    "httpx >= 0.27",
    "pydantic >= 2.0",
    "beautifulsoup4 >= 4.12",
    "lxml >= 5.0",
    "pymupdf4llm >= 0.0.17",
    "patchright >= 1.50",
    "chromadb >= 0.5.0",
    "platformdirs >= 4.0",
    "python-dotenv >= 1.0",
    "click >= 8.0",
    "rich >= 13.0",
]

[project.optional-dependencies]
marker = ["marker-pdf == 1.10.2"]
docling = ["docling"]
full = ["grados[marker,docling]"]

[project.scripts]
grados = "grados.cli:main"

[project.urls]
Homepage = "https://github.com/STSNaive/GRaDOS"
Repository = "https://github.com/STSNaive/GRaDOS"
Issues = "https://github.com/STSNaive/GRaDOS/issues"
Changelog = "https://github.com/STSNaive/GRaDOS/blob/main/CHANGELOG.md"

# ───── uv ─────

[dependency-groups]
dev = [
    "pytest >= 8.0",
    "ruff >= 0.4",
    "mypy >= 1.10",
]

# ───── tools ─────

[tool.ruff]
target-version = "py311"
line-length = 120

[tool.ruff.lint]
select = ["E", "F", "W", "I", "UP"]

[tool.mypy]
python_version = "3.11"
strict = true

[tool.pytest.ini_options]
filterwarnings = [
    "ignore:builtin type .* has no __module__ attribute:DeprecationWarning",
]

[tool.hatch.build.targets.wheel]
packages = ["src/grados"]

[tool.hatch.build.targets.sdist]
exclude = [
    ".uv-cache/",
    ".github/",
    ".claude/",
    ".grados/",
    "hooks/",
    "marker-worker/",
    "skills/",
    "docs/",
    "node_modules/",
    "dist/",
    ".mcp.json",
    "uv.lock",
    "grados-config.json",
    "grados-config.example.json",
    "*-plan.md",
    "*-decisions.md",
    "*-checklist.md",
    "status.md",
    "TODO.md",
]
